requirements.txt
djangoamqp==1.4.9
anyjson==0.3.3
AuthorizeSauce==0.4.1
billiard==3.3.0.23
celery==3.1.23
coreapi==2.3.3
coreschema==0.0.4
decorator==4.1.2
Django==1.9.2
django-braces==1.9.0
django-celery==3.1.17
django-cors-headers==2.1.0
django-custom-user==0.6
django-docs==0.1.2
django-dynamic-logging==1.0.3
django-environ==0.4.4
django-health-check==2.4.0
django-model-utils==2.5
django-rest-swagger==2.1.0
django-viewsets==0.1.9
django-webpack-loader==0.5.0
django-widget-tweaks==1.4.1
djangorestframework==3.5.3
djangorestframework-jwt==1.8.0
djoser==0.5.1
facebook-sdk==2.0.0
html5lib==1.0b8
ipython==6.2.1
ipython-genutils==0.2.0
itypes==1.1.0
jedi==0.11.0
Jinja2==2.9.6
jsonfield==1.0.3
kombu==3.0.37
MarkupSafe==1.0
oauthlib==2.0.6
olefile==0.44
openapi-codec==1.3.2
parso==0.1.0
pexpect==4.3.0
pickleshare==0.7.4
Pillow==4.3.0
prompt-toolkit==1.0.15
ptyprocess==0.5.2
py==1.4.34
Pygments==2.2.0
PyJWT==1.5.3
pymongo==3.2.2
pytest==3.0.5
python-http-client==1.2.3
pytz==2016.6.1
requests==2.10.0
requests-oauthlib==0.8.0
simplegeneric==0.8.1
simplejson==3.12.0
six==1.10.0
smtpapi==0.3.1
sqlparse==0.2.1
suds-jurko==0.6
traitlets==4.3.2
tweepy==3.5.0
uritemplate==3.0.0
wcwidth==0.1.7
